# improve performance of is_all_triangles