### Improve performance of ``is_all_triangles``.

Our ``is_all_triangles`` checks the ``faces`` array to determine if each face in ``PolyData`` is indeed a triangle.  However, when we check this, we rebuild the ``faces`` array each time, which in ``vtk>=9.0`` is rebuit within each time ``GetPolys`` is called.

To improve performance, we simply cache `faces` and then verify the shape and content:

### Test code:

```py
import pyvista as pv
from pyvista import examples
mesh = examples.download_nefertiti()
```

#### This branch

```py
>>> mesh.is_all_triangles()
# 18.5 ms ± 555 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

#### Master

```py
>>> mesh.is_all_triangles()
# 46.6 ms ± 103 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```